move domain query tests
