+++ api-js/src/queries/domains/__tests__/find-domain-by-domain.test.js
@@ -4,18 +4,18 @@ const { graphql, GraphQLSchema, GraphQLError } = require('graphql')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { createQuerySchema } = require('../queries')
-const { createMutationSchema } = require('../mutations')
-const { cleanseInput } = require('../validators')
-const { checkDomainPermission, tokenize, userRequired } = require('../auth')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { createQuerySchema } = require('../..')
+const { createMutationSchema } = require('../../../mutations')
+const { cleanseInput } = require('../../../validators')
+const { checkDomainPermission, tokenize, userRequired } = require('../../../auth')
 const {
   userLoaderByUserName,
   domainLoaderByDomain,
   userLoaderByKey,
-} = require('../loaders')
+} = require('../../../loaders')
 const { DB_PASS: rootPass, DB_URL: url } = process.env
 
 describe('given findDomainByDomain query', () => {
